requirements-minimal.txt
fastapi==0.135.3
uvicorn==0.43.0
groq==0.37.1
langchain-core==1.2.26
langchain-groq==1.1.2
pydantic==2.12.5
python-dotenv==1.2.2
requests==2.33.1
# Lightweight dependencies only - no heavy ML packages